# Manager: Fix panel reactivity