Closes https://github.com/storybookjs/storybook/issues/30340

<!-- If your PR is related to an issue, provide the number(s) above; if it resolves multiple issues, be sure to break them up (e.g. "closes #1000, closes #1001"). -->

<!--

Thank you for contributing to Storybook! Please submit all PRs to the `next` branch unless they are specific to the current release. Storybook maintainers cherry-pick bug and documentation fixes into the `main` branch as part of the release process, so you shouldn't need to worry about this. For additional guidance: https://storybook.js.org/docs/contribute

-->

## What I did

- Fixed memoization issues of the manager addon panel

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:

- [ ] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli-storybook/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

  - `bug`: Internal changes that fixes incorrect behavior.
  - `maintenance`: User-facing maintenance tasks.
  - `dependencies`: Upgrading (sometimes downgrading) dependencies.
  - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
  - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
  - `documentation`: Documentation **only** changes. Will not show up in release changelog.
  - `feature request`: Introducing a new feature.
  - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
  - `other`: Changes that don't fit in the above categories.

   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->
This pull request has been released as version `0.0.0-pr-30638-sha-e89b3763`. Try it out in a new sandbox by running `npx storybook@0.0.0-pr-30638-sha-e89b3763 sandbox` or in an existing project with `npx storybook@0.0.0-pr-30638-sha-e89b3763 upgrade`.
<details>
<summary>More information</summary>

| | |
| --- | --- |
| **Published version** | [`0.0.0-pr-30638-sha-e89b3763`](https://npmjs.com/package/storybook/v/0.0.0-pr-30638-sha-e89b3763) |
| **Triggered by** | @valentinpalkovic |
| **Repository** | [storybookjs/storybook](https://github.com/storybookjs/storybook) |
| **Branch** | [`valentin/fix-addon-panel-reactivity`](https://github.com/storybookjs/storybook/tree/valentin/fix-addon-panel-reactivity) |
| **Commit** | [`e89b3763`](https://github.com/storybookjs/storybook/commit/e89b376313a9cf032e0f0362e7598a397601dc0e) |
| **Datetime** | Sun Feb 23 19:53:09 UTC 2025 (`1740340389`) |
| **Workflow run** | [13486367981](https://github.com/storybookjs/storybook/actions/runs/13486367981) |

To request a new release of this pull request, mention the `@storybookjs/core` team.

_core team members can create a new canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=30638`_
</details>
<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  80.6 MB | 80.6 MB | 0 B | **1.81** | 0% |
| initSize |  80.6 MB | 80.6 MB | 0 B | **1.81** | 0% |
| diffSize |  97 B | 97 B | 0 B | - | 0% |
| buildSize |  7.31 MB | 7.32 MB | 4.69 kB | **24.39** | 0.1% |
| buildSbAddonsSize |  1.9 MB | 1.9 MB | 0 B | -0.26 | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  1.88 MB | 1.88 MB | 4.69 kB | **17.43** | 0.2% |
| buildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  3.97 MB | 3.97 MB | 4.69 kB | **27.76** | 0.1% |
| buildPreviewSize |  3.34 MB | 3.34 MB | 0 B | 0.62 | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  6.6s | 7.6s | 956ms | -0.64 | 12.5% |
| generateTime |  17s | 17.6s | 532ms | -1.13 | 3% |
| initTime |  4.2s | 4s | -251ms | **-1.65** | 🔰-6.3% |
| buildTime |  9.1s | 8.6s | -461ms | -0.73 | -5.3% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  5s | 5.3s | 298ms | -0.11 | 5.5% |
| devManagerResponsive |  4.8s | 5s | 228ms | 0.12 | 4.5% |
| devManagerHeaderVisible |  776ms | 739ms | -37ms | -0.5 | -5% |
| devManagerIndexVisible |  789ms | 766ms | -23ms | -0.54 | -3% |
| devStoryVisibleUncached |  1.7s | 2s | 300ms | -0.31 | 14.7% |
| devStoryVisible |  816ms | 773ms | -43ms | -0.78 | -5.6% |
| devAutodocsVisible |  681ms | 764ms | 83ms | -0.07 | 10.9% |
| devMDXVisible |  761ms | 646ms | -115ms | **-1.56** | 🔰-17.8% |
| buildManagerHeaderVisible |  876ms | 660ms | -216ms | -0.82 | -32.7% |
| buildManagerIndexVisible |  888ms | 754ms | -134ms | -0.36 | -17.8% |
| buildStoryVisible |  830ms | 640ms | -190ms | -0.78 | -29.7% |
| buildAutodocsVisible |  758ms | 682ms | -76ms | 0.79 | -11.1% |
| buildMDXVisible |  563ms | 494ms | -69ms | -1.19 | -14% |

<!-- BENCHMARK_SECTION -->


<!-- greptile_comment -->

## Greptile Summary

Refactored the Panel component in Storybook's manager to improve reactivity by replacing the Consumer pattern with React hooks, fixing the issue of controls not showing on first load.

- Replaced Consumer pattern with `useStorybookApi` and `useStorybookState` hooks in `code/core/src/manager/container/Panel.tsx`
- Added story state tracking through `STORY_PREPARED` event using `useChannel` hook
- Optimized panel filtering logic with `useMemo` hooks
- Removed dependency on memoizerific library



<!-- /greptile_comment -->